Get screen clearing working
